refactor(neon): Make account selection reusable
